28/12/2023
i've spent the last few weeks scared out of my mind. for both of us! and all of that time i've been trying to...manage you. you're happy as larry one moment, furious or upset the next. i've heard you scream in the night. i've heard you crying in the mornings- do you think this is fun for me?